Align the alarm timer to the next minute
This doesn't really change the semantics, but it makes things appear much
more responsive/reliable as alarms now pop up the second your clock
switches to the next minute.

While at it this also no longer spins the alarm timer in case Akonadi isn't
set up yet and removes a few config options that were never written to and
that unnecessarily complicated this.