Fold NotificationHandler into KalendarAlarmClient
Those two have a 1:1 relation and the separation doesn't really give us
any value but requires additional code to tie things together.

The intended effect of this is that we now have access to the calendar
object when creating notifications, which is necessary for moving beyond
a single string and provide more detailed notifications eventually.