Drag and drop convert to and from all day events
In the week view it is now possible to drag and drop allday/multiday
event to the "main area" to convert it to be 1h long event.
Drag and drop in the opposite way also works now. So it is possible
to drag and drop hourly event to the allday/multiday events area to
convert it to be all day task.

Closes-issue: #120