Determine occurrence of recurring incidences for an alarm
The main usecase for this is opening recurring events on the right
day in the calendar.

This is based on the same logic that Calendar::appendRecurringAlarms()
uses, but with ignoring two cases:
- alarms at fixed times rather than relative offsets
- recurring alarms

Handling of those cases can be added as well if needed, but those are
rare and considerably complicate the code. As we only need the
occurrence for "comfort features" rather than correct alarm displays
that seems acceptable for now.