Do not require leading "0" for hours while creating event
BUG: 465674

Issue : When creating or editing an event, Kalendar currently requires to input the time in full four-digit format. This means that "8" or "5" are not accepted as valid input in the hour section, therefore requiring the user to input a leading 0 ("08", "05"). There is no indication whatsoever that this is the issue, though, so the user just finds themselves with the input being reset without any explanation.

Solution: Treat single-digit input in the hours section as having a leading zero by default