Fix crashes in incidenceoccurrencemodel
The cause for these crashes is that the QPersistentModelIndex get
invalided at some point and then the row is -1 which is invalid.
This is probably caused by bugs somewhere in ETMCalendar.

The previous code which tried to only update part of the model,
which in theory should be faster was wrong since a dataChanged on the
source model could result in less or more entry due to the recurrence
rules changing or in a different hash due to the start or end date
changing.

BUG: 466063


(cherry picked from commit c925a526ceb89d78514ead777576987b620cc516)